Fix 'unknown unknowns' being highlighted on first page load
Scrollspy was being initialized twice, once with data attributes in an
inner div, once with javascript in the body. The docs suggest it should
be in the body, so this removes the javascript and moves the data
attributes to the body.

This also removes data-offset="0" to use the default value of 10, which
is needed when using the body. It also adds 'position: relative' to the
body because the docs say it's required, but it doesn't seem to do
anything in my browser.